+++ src/functions/project/init-terraform.ts
@@ -1,4 +1,3 @@
-import { WorkspaceContext } from '@causa/workspace';
 import { ProjectInit } from '@causa/workspace-core';
 import { rm } from 'fs/promises';
 import { join } from 'path';
@@ -13,30 +12,33 @@ const TERRAFORM_DIR = '.terraform';
  * Implements the {@link ProjectInit} function for Terraform projects, by running `terraform init`.
  */
 export class ProjectInitForTerraform extends ProjectInit {
-  async _call(context: WorkspaceContext): Promise<void> {
-    const projectName = context.get('project.name');
+  async _call(): Promise<void> {
+    const projectName = this._context.get('project.name');
 
     if (this.force) {
-      context.logger.info('🔥 Removing Terraform folder.');
+      this._context.logger.info('🔥 Removing Terraform folder.');
 
-      const terraformDir = join(context.getProjectPathOrThrow(), TERRAFORM_DIR);
+      const terraformDir = join(
+        this._context.getProjectPathOrThrow(),
+        TERRAFORM_DIR,
+      );
       await rm(terraformDir, { recursive: true, force: true });
     }
 
-    context.logger.info(
+    this._context.logger.info(
       `️🎉 Initializing Terraform for project '${projectName}'.`,
     );
 
-    await context.service(TerraformService).init({ logging: 'debug' });
+    await this._context.service(TerraformService).init({ logging: 'debug' });
 
-    context.logger.info(`️✅ Successfully initialized Terraform.`);
+    this._context.logger.info(`️✅ Successfully initialized Terraform.`);
   }
 
-  _supports(context: WorkspaceContext): boolean {
+  _supports(): boolean {
     return (
       !this.workspace &&
-      context.get('project.language') === 'terraform' &&
-      context.get('project.type') === 'infrastructure'
+      this._context.get('project.language') === 'terraform' &&
+      this._context.get('project.type') === 'infrastructure'
     );
   }
 }